Refactored the diffParser to be more adequately commented, and also re-ordered the methods some to reflect when they are used. Also cleaned up some things in general.
