+++ cli/gitlit.js
@@ -6,21 +6,17 @@ var init = require("./repoInit");
 var writer = require("./htmlWriter");
 var program = require("commander");
 
-function printExtraHelp(){
-    console.log('IMPORTANT NOTE:');
-    console.log('gitlit uses the parse5 HTML parsing library,\n' + '' +
-        'as such, there is undefined behavior if there are missing\n' +
-        'opening or closing tags. If undesired behavior occurs,\n' +
-        'check your document for missing opening or closing tags.');
-}
-
+// Set commander info that will display during the help
 program._name = 'gitlit';
 
 program
-    .version('0.3.1')
+    .version('0.5.0')
     .option('-v, --versionFull', 'Print out all the version info for the CLI')
     .option('-l, --libraries', 'Print out the versions of the libraries used');
 
+/*
+ * This section set the commands, their signatures, and what they do
+ */
 program
     .command('init <file> <outputPath> <repoName>')
     .description('\n\tInitialize a Repository for the given file')
@@ -48,15 +44,28 @@ program
     .description('\n\tShow the difference between the last 2 revisions of the repository'
                 + '\n\tDestination should be the directory the output should end up')
     .action(function(repoLocation, outputLocation) {
+        // Get the file before and after changes
         var fileVersions = init.getOldAndNewFileVersions(repoLocation);
         var mergeFileVersions = init.getOldAndNewFileVersions(repoLocation);
-        var changes = init.getInterprettedDiff(repoLocation);
+
+        // Get the interpreted diff that tells what the differences are in more detail
+        var changes = init.getInterpretedDiff(repoLocation);
+
+        //Create pairs that match up what should go side-by-side
         var diffDisplayPairs = init.createDiffPairs(fileVersions[0], fileVersions[1], changes);
-        init.matchMovesEdits(diffDisplayPairs);
-        var diffDisplayObject = init.setUpPairsForDiffDisplay(diffDisplayPairs);
         var mergePairs = init.createMergePairs(mergeFileVersions[0], mergeFileVersions[1], changes);
-        init.matchMovesEdits(mergePairs);
+
+        //Label any moves and edits only once so that only one button appears for each actual change
+        init.labelUniqueMovesAndEdits(diffDisplayPairs);
+        init.labelUniqueMovesAndEdits(mergePairs);
+
+        // Set up the pairs to be rendered
+        var diffDisplayObject = init.setUpPairsForDiffDisplay(diffDisplayPairs);
+
+        //Copy the files that actually allow for viewing of the diff
         init.createCopyOfDiffResources(outputLocation);
+
+        //Create the file that contains the information to be used in the diff rendering
         init.createJSONForDisplay(outputLocation, diffDisplayObject, mergePairs, mergeFileVersions[1]);
     });
 
@@ -77,6 +86,16 @@ if (process.argv.length == 2){
     printExtraHelp();
 }
 
+/*
+ * Print the warning about parse5 possibly interpreting bad HTML
+ */
+function printExtraHelp(){
+    console.log('IMPORTANT NOTE:');
+    console.log('gitlit uses the parse5 HTML parsing library,\n' + '' +
+        'as such, there is undefined behavior if there are missing\n' +
+        'opening or closing tags. If undesired behavior occurs,\n' +
+        'check your document for missing opening or closing tags.');
+}
 
 program.on('--help', function(){
     printExtraHelp();
